update ship's health

+++ object.cpp
@@ -1,12 +1,12 @@
 #include "object.h"
 
-void Object::setDest(int x, int y, int w, int h) {
+void Object::setDest(double x, double y, double w, double h) {
   dest.x=x;
   dest.y=y;
   dest.w=w;
   dest.h=h;
 }
-void Object::setSource(int x, int y, int w, int h) {
+void Object::setSource(double x, double y, double w, double h) {
   src.x=x;
   src.y=y;
   src.w=w;
@@ -15,4 +15,5 @@ void Object::setSource(int x, int y, int w, int h) {
 void Object::setImage(string filename, SDL_Renderer* ren) {
   SDL_Surface* surf = IMG_Load(filename.c_str());
   tex = SDL_CreateTextureFromSurface(ren, surf);
+  //  SDL_SetTextureBlendMode(tex, SDL_BLENDMODE_NONE);
 }
