+++ object.h
@@ -16,8 +16,8 @@ public:
 Object() {}
 SDL_Rect getDest() const {return dest;}
 SDL_Rect getSource() const {return src;}
-void setDest(int x, int y, int w, int h);
-void setSource(int x, int y, int w, int h);
+void setDest(double x, double y, double w, double h);
+void setSource(double x, double y, double w, double h);
 void setImage(string filename, SDL_Renderer* ren);
 SDL_Texture* getTex() const {return tex;}
 };
